requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pymemcache==3.4.0
python-dotenv==0.15.0
six==1.15.0
Werkzeug==1.0.1
